requirements.txt
pyside6
qt-material
event-bus
jsonpickle
numpy-stl
matplotlib
qtawesome